feat: add AudioPlaybackView component for audio playback in detail and form page
