+++ thoughtful/Localizable.xcstrings
@@ -149,9 +149,6 @@
     "Play" : {
 
     },
-    "Play/Stop" : {
-
-    },
     "Previous Week" : {
 
     },
