+++ thoughtful/Views/Thoughts/ThoughtDetailView.swift
@@ -11,9 +11,6 @@ import SwiftUI
 struct ThoughtDetailView: View {
     @State var thoughtVm: ThoughtViewModel = .init()
 
-    @StateObject private var player: AudioPlayer = .init()
-    @State private var timeRemaining: TimeInterval = 0
-
     @ObservedObject var thought: Thought
     @Environment(\.dismiss) var dismiss
     @Environment(\.modelContext) var context
@@ -28,58 +25,6 @@ struct ThoughtDetailView: View {
     }
 
     @State var photo: UIImage? = nil
-
-    private func playAudio() {
-        guard let fileName = thought.audioFileName else {
-            print("No audio")
-            return
-        }
-
-        let url = AudioRecorder.getFileURL().appendingPathComponent(fileName)
-
-        do {
-            let isReachable = try url.checkResourceIsReachable()
-
-            if !isReachable {
-                print("URL is not reachable")
-                return
-            }
-
-            print("URL is reachable: \(isReachable)")
-
-            print("Playing audio: \(url)")
-            player.play(url)
-        } catch {
-            print("Can't play thought audio: \(error)")
-        }
-    }
-
-    private func stopAudio() {
-        player.stop()
-    }
-
-    private func startTimer() {
-        if let duration = thought.audioDuration {
-            timeRemaining = duration
-            Timer.scheduledTimer(withTimeInterval: 1.0, repeats: true) { timer in
-                if timeRemaining > 0 {
-                    timeRemaining -= 1
-                } else {
-                    timer.invalidate()
-                }
-            }
-        }
-    }
-
-    private func resetTimer() {
-        timeRemaining = 0
-    }
-
-    private func formatTime(_ time: TimeInterval) -> String {
-        let minutes = Int(time) / 60
-        let seconds = Int(time) % 60
-        return String(format: "%02d:%02d", minutes, seconds)
-    }
 
     var body: some View {
         ScrollView {
@@ -112,31 +57,7 @@ struct ThoughtDetailView: View {
                         .font(.caption2)
                         .foregroundStyle(.primary.opacity(0.5))
 
-                    HStack(alignment: .center) {
-                        Button {
-                            if player.isPlaying {
-                                stopAudio()
-                                return
-                            }
-
-                            playAudio()
-                        } label: {
-                            if player.isPlaying {
-                                Label("Stop", systemImage: "stop.fill")
-                                    .labelStyle(.iconOnly)
-                            } else {
-                                Label("Play", systemImage: "play.fill")
-                                    .labelStyle(.iconOnly)
-                            }
-                        }
-                        .frame(width: 40, height: 40)
-                        .background(Color.accentColor, in: RoundedRectangle(cornerRadius: 24))
-
-                        Spacer()
-
-                        Text(formatTime(timeRemaining))
-                            .font(.caption)
-                    }
+                    AudioPlaybackView(thought: thought)
                 }
 
                 if thought.emotionExists {
@@ -202,11 +123,6 @@ struct ThoughtDetailView: View {
         .onAppear {
             thoughtVm.context = context
         }
-        .onAppear {
-            if let duration = thought.audioDuration {
-                timeRemaining = duration
-            }
-        }
         .task {
             DispatchQueue.global(qos: .background).async {
                 guard let photo = thought.photo else {
@@ -218,17 +134,6 @@ struct ThoughtDetailView: View {
                 withAnimation {
                     self.photo = loadedPhoto
                 }
-            }
-        }
-        .onChange(of: player.isPlaying) { oldValue, newValue in
-            if oldValue == false, newValue == true {
-                startTimer()
-                return
-            }
-
-            if oldValue == true, newValue == false {
-                resetTimer()
-                return
             }
         }
         .onChange(of: thought.photos) { _, photos in
@@ -270,12 +175,6 @@ struct ThoughtDetailView: View {
         }
     }
 }
-
-// extension ThoughtDetailView: AVAudioPlayerDelegate {
-//    func audioPlayerDidFinishPlaying(_ player: AVAudioPlayer, successfully flag: Bool) {
-//        isPlaying = false
-//    }
-// }
 
 #Preview {
     NavigationStack {
