+++ thoughtful/Views/Toolbar/ToolbarView.swift
@@ -64,20 +64,20 @@ struct ToolbarView: View {
             }
             .frame(maxWidth: .infinity)
 
-            Button {
-                if player.isPlaying {
-                    player.stop()
-                    return
-                }
-
-                let url = recorder.filePath
-                player.play(url!)
-            } label: {
-                Label("Play/Stop", systemImage: player.isPlaying ? "stop.fill" : "play.fill")
-                    .labelStyle(.iconOnly)
-            }
-            .disabled(recorder.filePath == nil)
-            .frame(maxWidth: .infinity)
+//            Button {
+//                if player.isPlaying {
+//                    player.stop()
+//                    return
+//                }
+//
+//                let url = recorder.filePath
+//                player.play(url!)
+//            } label: {
+//                Label("Play/Stop", systemImage: player.isPlaying ? "stop.fill" : "play.fill")
+//                    .labelStyle(.iconOnly)
+//            }
+//            .disabled(recorder.filePath == nil)
+//            .frame(maxWidth: .infinity)
 
             PhotosPicker(selection: $photosPickerItem, matching: .images) {
                 Label("Open Photos", systemImage: "photo.fill")
